requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
tensorflow>=2.6.0
PyQt5>=5.15.4
pefile>=2021.5.24
yara-python>=4.2.3
requests>=2.26.0
python-magic-bin>=0.4.14
joblib>=1.0.1
psutil>=5.8.0
pillow>=10.0.0 